changed packetNo to use int everywhere, it sometimes was long.
packetNo can be an int, which limits file size to 403.726.925.636 bytes, or 375 GB. Should be enough for everybody. (Famous last words)

+++ src/main/java/nl/digitalekabeltelevisie/data/mpeg/TransportStream.java
@@ -237,7 +237,7 @@ public class TransportStream implements TreeNode{
 	public void parseStream(final java.awt.Component component) throws IOException {
 		final PositionPushbackInputStream fileStream = getInputStream(component);
 		final byte [] buf = new byte[packetLength];
-		long count=0;
+		int count=0;
 		no_packets = 0;
 
 		pids = new PID [8192];
@@ -323,7 +323,7 @@ public class TransportStream implements TreeNode{
 		}
 		final PositionPushbackInputStream fileStream = getInputStream(component);
 		final byte [] buf = new byte[packetLength];
-		long count=0;
+		int count=0;
 
 
 		int bytes_read =0;
